requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.3
jiter==0.8.2
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.2.1
numba==0.61.0
numpy==1.26.3
openai==1.59.9
pillow==10.2.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pypiwin32==223
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
srt==3.5.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
vosk==0.3.45
websockets==14.2
